+++ src/api/index.js
@@ -1,4 +1,4 @@
-export { createStore } from './store';
+export { createStore, persisted } from './store';
 export { createHistory } from './history';
 export { default as addMiddlewares } from './middlewares';
 export { createModel, createVirtualModel } from './configure';