+++ src/pubsub/events.js
@@ -1,14 +1,14 @@
 import { createDefault } from 'utils';
 
-import { needLogs } from './utils';
+import { log } from './logger';
 
 const events = {};
 
 const doesEventExist = event => !!events[event];
 
 const createEvent = (event, inital, options) => {
-  events[event] = createDefault(inital, options);
-  needLogs && console.log(`Event has been created: ${event}.`);
+  events[event] = createDefault({ lastState: inital, options });
+  log(`Event has been created: ${event}.`, false);
   return getEvent(event);
 }